// modules/doctor/constants/enums/approach.enum.ts
export enum ApproachEnum {
  SEM_PREFERENCIA = 'sem_preferencia',
  COMPORTAMENTAL_COGNITIVO = 'comportamental_cognitivo',
  HUMANISTA = 'humanista',
  ANALISE_COMPORTAMENTO = 'analise_comportamento',
  PSICANALISE = 'psicanalise',
  GESTALT = 'gestalt',
  FENOMENOLOGICA_EXISTENCIAL = 'fenomenologica_existencial',
  COMPORTAMENTAL_CONTEXTUAL = 'comportamental_contextual',
  OUTRO = 'outro',
}